fix return code of remove, add more unit-tests
